requirements.txt
requests>=2.28
python-dotenv>=1.0
pypdf>=5.0
pytest>=8.0
